feat(seo): standalone /tarifs route with Product/Offer/FAQ JSON-LD
- Extract PricingSection out of routes/index.tsx into
  components/landing/pricing-section.tsx so it can be reused
- New routes/tarifs.tsx: hero + reused PricingSection + 6-question FAQ.
  SoftwareApplication is on the landing JSON-LD; this page emits a
  Product with three Offer nodes (Free, Famille annual, Famille monthly,
  with billingDuration P1Y/P1M) plus FAQPage for featured snippets
- Replace landing nav anchors (desktop + mobile sheet) and ressources
  TopNav 'Tarifs' from #tarifs / hash='tarifs' to /tarifs route
- robots.txt + sitemap.xml: add /tarifs
