+++ apps/web/public/sitemap.xml
@@ -11,6 +11,11 @@
     <priority>0.5</priority>
   </url>
   <url>
+    <loc>https://toko.battistella.ovh/tarifs</loc>
+    <changefreq>monthly</changefreq>
+    <priority>0.9</priority>
+  </url>
+  <url>
     <loc>https://toko.battistella.ovh/contact</loc>
     <changefreq>monthly</changefreq>
     <priority>0.5</priority>
